move all logic for the touchPoints useReducer to store.js
